feat: Adding Types for shopify.
Adding required shopify products types.
Author: ManojPanda3
SignedBy: ManojPanda3
